Refactor imports and restructure components; update metadata in HTML and manifest
