+++ src/index.tsx
@@ -1,18 +1,18 @@
-import { StoreProvider } from 'Components/Context/store'
-import { Workspace } from 'Components/Workspace'
+import { StoreProvider } from 'store/store'
+import { Workspace } from 'pages/Workspace/Workspace'
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import { createBrowserRouter, Outlet, RouterProvider } from 'react-router-dom'
-import { ROUTE } from 'services/routes'
-import { applyTheme, getInitialTheme, watchSystemTheme } from 'services/theme'
+import { ROUTE } from 'routes'
+import { applyTheme, getInitialTheme, watchSystemTheme } from 'lib/theme'
 
-import { Account } from './Components/Account'
-import { Auth } from './Components/Auth'
-import { Layout } from './Components/Layout'
-import { Privacy } from './Components/Legal/Privacy'
-import { Terms } from './Components/Legal/Terms'
-import { NotFound } from './Components/NotFound'
-import { TaskPreview } from './Components/TaskPreview'
+import { Account } from 'pages/Account/Account'
+import { Auth } from 'pages/Auth'
+import { Layout } from 'components/layout/Layout'
+import { Privacy } from 'pages/Legal/Privacy'
+import { Terms } from 'pages/Legal/Terms'
+import { NotFound } from 'pages/NotFound'
+import { TaskPreview } from 'pages/TaskPreview/TaskPreview'
 import './index.css'
 
 const container = document.getElementById('wunderTodo')
