+++ src/pages/Legal/Terms.tsx
