+++ src/routes.ts
@@ -1,0 +1,9 @@
+enum ROUTE {
+  ROOT = '/',
+  TASK_PAGE = '/t/:id',
+  TERMS = '/terms',
+  PRIVACY = '/privacy',
+  ACCOUNT = '/account',
+}
+
+export { ROUTE }
